add comments; remove cruft; reorganize files
